+++ rl_blox/algorithm/td7.py
@@ -11,155 +11,17 @@ import tqdm
 from flax import nnx
 
 from ..blox.double_qnet import ContinuousClippedDoubleQNet
+from ..blox.embedding.sale import (
+    SALE,
+    ActorSALE,
+    CriticSALE,
+    state_action_embedding_loss,
+)
 from ..blox.function_approximator.mlp import MLP
-from ..blox.function_approximator.norm import avg_l1_norm
 from ..blox.function_approximator.policy_head import DeterministicTanhPolicy
 from ..blox.replay_buffer import LAP
 from ..blox.target_net import hard_target_net_update
 from ..logging.logger import LoggerBase
-
-
-class SALE(nnx.Module):
-    r"""SALE: state-action learned embedding.
-
-    Although the embeddings are learned by considering the dynamics of the
-    environment, their purpose is solely to improve the input to the value
-    function and policy, and not to serve as a world model for planning or
-    estimating rollouts.
-
-    Parameters
-    ----------
-    state_embedding : nnx.Module
-        State embedding network **without** AvgL1Norm. AvgL1Norm will be added
-        in this module to form :math:`z^s = f(s)`. This networks maps state to
-        **unnormalized** zs.
-
-    state_action_embedding : nnx.Module
-        :math:`z^{sa} = g(z^s, a)`. State action embedding. Maps zs and action
-        to zsa, which is trained to be the same as the normalized zs of the
-        next state.
-    """
-
-    _state_embedding: nnx.Module
-    state_action_embedding: nnx.Module
-
-    def __init__(
-        self, state_embedding: nnx.Module, state_action_embedding: nnx.Module
-    ):
-        self.state_embedding = state_embedding
-        self.state_action_embedding = state_action_embedding
-
-    def __call__(
-        self, state: jnp.ndarray, action: jnp.ndarray
-    ) -> tuple[jnp.ndarray, jnp.ndarray]:
-        zs = self.state_embedding(state)
-        zs_action = jnp.concatenate((zs, action), axis=-1)
-        zsa = self.state_action_embedding(zs_action)
-        return zsa, zs
-
-    def state_embedding(self, state: jnp.ndarray) -> jnp.ndarray:
-        r""":math:`z^s = f(s)`."""
-        return avg_l1_norm(self._state_embedding(state))
-
-
-class ActorSALE(nnx.Module):
-    """Deterministic policy with SALE.
-
-    The actor maps a state vector through a linear leayer and applied AvgL1Norm,
-    it concatenates the result with the embedded vector zs, and then maps it
-    through a deterministic policy net to the action.
-
-    This module implements the function :math:`\pi(s, z^s_t)`.
-
-    Parameters
-    ----------
-    policy_net : nnx.Module
-        Deterministic policy network.
-
-    n_state_features : int
-        Number of state components.
-
-    hidden_nodes : int
-        Number of nodes in the first layer that encodes state and applies
-        :func:`~.blox.function_approximator.norm.avg_l1_norm`.
-
-    rngs : nnx.Rngs
-        Random number generator.
-    """
-
-    policy_net: nnx.Module
-    l0: nnx.Linear
-
-    def __init__(
-        self,
-        policy_net: nnx.Module,
-        n_state_features: int,
-        hidden_nodes: int,
-        rngs: nnx.Rngs,
-    ):
-        self.policy_net = policy_net
-        self.l0 = nnx.Linear(n_state_features, hidden_nodes, rngs=rngs)
-
-    def __call__(self, state: jnp.ndarray, zs: jnp.ndarray) -> jnp.ndarray:
-        """pi(state, zs)."""
-        h = avg_l1_norm(self.l0(state))
-        he = jnp.concatenate((h, zs), axis=-1)
-        return self.policy_net(he)
-
-
-class CriticSALE(nnx.Module):
-    r"""Action-value function Q with SALE.
-
-    The critic maps a concatenated state-action vector through a linear layer
-    and applies AvgL1Norm, it concatenates the result with the embedded vectors
-    zs and zsa of the state-action pair, and then maps it through the q_net
-    to the expected value of the state-action pair.
-
-    This module implements the function :math:`Q(s, a, z^{sa}_t z^s_t)`.
-
-    Parameters
-    ----------
-    q_net : nnx.Module
-        Action-value function.
-
-    n_state_features : int
-        Number of state components.
-
-    n_action_features : int
-        Number of action components.
-
-    hidden_nodes : int
-        Number of nodes in the first layer that encodes state and action and
-        applies :func:`~.blox.function_approximator.norm.avg_l1_norm`.
-
-    rngs : nnx.Rngs
-        Random number generator.
-    """
-
-    q_net: nnx.Module
-    q0: nnx.Linear
-
-    def __init__(
-        self,
-        q_net: nnx.Module,
-        n_state_features: int,
-        n_action_features: int,
-        hidden_nodes: int,
-        rngs: nnx.Rngs,
-    ):
-        self.q_net = q_net
-        self.q0 = nnx.Linear(
-            n_state_features + n_action_features, hidden_nodes, rngs=rngs
-        )
-
-    def __call__(
-        self, sa: jnp.ndarray, zsa: jnp.ndarray, zs: jnp.ndarray
-    ) -> jnp.ndarray:
-        """Q(s, a, zsa, zs)."""
-        h = avg_l1_norm(self.q0(sa))
-        embeddings = jnp.concatenate((zsa, zs), axis=-1)
-        he = jnp.concatenate((h, embeddings), axis=-1)
-        return self.q_net(he)
 
 
 @dataclasses.dataclass
@@ -177,54 +39,6 @@ class CheckpointState:
     max_episodes_before_update: int = 1
     min_return: float = 1e8
     best_min_return: float = -1e8
-
-
-def state_action_embedding_loss(
-    embedding: SALE,
-    observation: jnp.ndarray,
-    action: jnp.ndarray,
-    next_observation: jnp.ndarray,
-) -> float:
-    r"""Loss of state-action embedding.
-
-    The encoders are jointly trained using the mean squared error (MSE) between
-    the state-action embedding :math:`z^{sa}` and the embedding of the next
-    state :math:`z^{s'}:
-
-    .. math:
-
-        \mathcal{L} = \frac{1}{N} \sum_i \left(
-        z^{s_i,a_i} - \texttt{sg}(z^{s_i'}) \right)^2,
-
-    where :math:`\texttt{sg}(\cdot)` denotes the stop-gradient operation.
-
-    The embeddings are designed to model the underlying structure of the
-    environment. However, they may not encompass all relevant information
-    needed by the value function and policy, such as features related to the
-    reward, current policy, or task horizon.
-
-    Parameters
-    ----------
-    embedding : SALE
-        State-action learned embedding.
-
-    observation : array, shape (batch_size,) + observation_space.shape
-        Observations.
-
-    action : array, shape (batch_size,) + action.shape
-        Action.
-
-    next_observation : array, shape (batch_size,) + observation_space.shape
-        Next observations.
-
-    Returns
-    -------
-    loss : float
-        Loss value.
-    """
-    zsa, _ = embedding(observation, action)
-    zsp = jax.lax.stop_gradient(embedding.state_embedding(next_observation))
-    return optax.squared_error(predictions=zsa, targets=zsp).mean()
 
 
 def sample_actions(
